Updates to SMF IGMP
Converted to using the json output commands for more stable parsing.
Added managed memberships to interfaces to control forwarding to host
interfaces. An interface is marked as managed when it is configured in FRR
with IGMP support, when marked as managed, only packets to groups
that have an active listener will be forwarded on that host interface.
Managed interfaces are automatically detected via FRR, but still require
the push/rpush command line options for them to be initially set up in SMF.

Signed-off-by: Nathan Bahr <nbahr@atcorp.com>